Elaborando as requisiçoes

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,9 +18,10 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marca = Marca::all();
+        // $marca = Marca::all();
+        $marca = $this->marca->all();
 
-        return $marca;
+        return response()->json($marca,200);
     }
 
    
@@ -30,44 +35,63 @@ class MarcaController extends Controller
     {
         // return 'Chegamos ate aqui';
         //dd($request->all());//Para debugar somente com o request
-        $marca = Marca::create($request->all());//Desta forma criamos o registro de forma massiva
-        return $marca;
+        // $marca = Marca::create($request->all());//Desta fkorma criamos o registro de forma massiva
+        $marca = $this->marca->create($request->all());
+        
+        return response()->json($marca,201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $marca foi troca do model para integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        //Aqui no show ele recebe o id da rota e automaticamente pesquisa em Marca o id recebido e retorna os dados;
-        return $marca;
+        $marca = $this->marca->find($id);
+
+        if($marca === null)
+            return response()->json(['erro' => 'Recurso Indisponivel para visualização'],404);
+     
+        return response()->json($marca,200);
     }
 
         /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $marca
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+
+        $marca = $this->marca->find($id);
+
+        if($marca === null)
+            return response()->json(['erro' => 'Recurso Indisponivel para atualização'],404);
+
         $marca->update($request->all());
-        return $marca;
+     
+        return response()->json($marca,200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $marca
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if($marca === null)
+            return response()->json(['erro' => 'Recurso Indisponivel para remoção'],404);
+
         $marca->delete();
-        return ['msg' => 'O registro foi removido do banco de dados'];
+     
+        return reponse()->json(['msg' => 'O registro foi removido do banco de dados'],200);
     }
 }
